src/requirements.txt
Flask
pymongo
git+https://github.com/leherreras/flask-negotiation.git
pyyaml
Flask-Classy
Flask-Paginate